requirements.txt
requests
bs4
pyyaml
pika
pymongo
python-dotenv
playwright
